backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
python-jose[cryptography]
passlib[argon2]
pydantic-settings
email-validator
python-dotenv
pytest
httpx
